mrbNG 5.00.00 with cetmodules integration.
* Builds and packages itself with cetmodules.

* Reorganize and separate external scripts from internal ones.

* Handle checkout and build of packages using cetmodules.

* Follow best practice for libraries in nested subprojects per
  https://gist.github.com/mbinna/c61dbb39bca0e4fb7d1f73b0d66a4fd1#using-a-library-defined-in-the-same-cmake-tree-should-look-the-same-as-using-an-external-library.

* copy_files_to_srcs.sh is now less destructive.

* parse_deps.pm is now imported from cetmodules when mrb is packaged
  rather than being managed as a fork.

* Complete rewrite of bin/setup_products to share more code with
  set_dev_products in cetmodules, and to interact appropriately with
  buildtools.

* Deactivate omnibus caching to avoid confusion with standalone
  cetmodules operation. CPack cannot easily produce separate archives
  for submodules.

+++ libexec/copy_files_to_srcs.sh
@@ -11,10 +11,15 @@ fullCom="${mrb_command} $thisCom"
 function usage() 
 {
   cat 1>&2 << EOF
-Usage: $fullCom <MRB_SOURCE>
+Usage: $fullCom [-c] <MRB_SOURCE>
   This is a utility function used by newDev and updateDepsCM.
 EOF
 }
+
+if [ "$1" = "-c" ]; then
+  (( clean = 1 ))
+  shift
+fi
 
 MRB_SOURCE=${1}
 
@@ -35,10 +40,17 @@ cmake_subdir_fragment=${MRB_SOURCE}/.cmake_add_subdir
 cmake_include_fragment=${MRB_SOURCE}/.cmake_include_dirs
 
 # Construct a new CMakeLists.txt file in srcs
-cp ${MRB_DIR}/templates/CMakeLists.txt.main ${MRB_SOURCE}/CMakeLists.txt || exit 1;
-if [ -e ${cmake_subdir_fragment} ]; then rm -f ${cmake_subdir_fragment}; fi
-if [ -e ${cmake_include_fragment} ]; then rm -f ${cmake_include_fragment}; fi
-echo "# DO NOT DELETE " > ${cmake_include_fragment}
-echo "# DO NOT DELETE " > ${cmake_subdir_fragment}
+mrb_dot_version=${MRB_VERSION#v}
+mrb_dot_version=${mrb_dot_version//_/.}
+sed -Ee 's&\@mrb_project\@&'"$MRB_PROJECT"'_MRB&' \
+  -e 's&\@mrb_dot_version\@&'"$mrb_dot_version"'&' \
+  $mrb_templates/CMakeLists.txt.main > ${MRB_SOURCE}/CMakeLists.txt || exit
+
+if (( clean )); then # Clear subdirectories.
+  if [ -e ${cmake_subdir_fragment} ]; then rm -f ${cmake_subdir_fragment}; fi
+  if [ -e ${cmake_include_fragment} ]; then rm -f ${cmake_include_fragment}; fi
+  echo "# DO NOT DELETE " > ${cmake_include_fragment}
+  echo "# DO NOT DELETE " > ${cmake_subdir_fragment}
+fi
 
 exit 0
